+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **1.2.61**
 - Added Multimon-based EAS decoder by Matthew McDougal.
 - Added separate, selectable bandplans for ITU regions.
+- Added Eclipse waterfall theme by Dimitar Milkov.
 - Added HFGCS, MARS, and CAF aviation bookmarks.
 - Sorting bands by low_band for in-order rendering.